farm value changed from int to long

+++ data/src/main/java/com/example/data/dto/PhotoItemDto.kt
@@ -9,5 +9,5 @@ data class PhotoItemDto(
     val owner: String,
     val secret: String,
     val server: String,
-    val farm: Int,
+    val farm: Long,
 )
