+++ data/src/main/java/com/example/data/network/model/PhotoSearchResponse.kt
@@ -20,7 +20,7 @@ internal data class Photo(
     @SerializedName("owner") val owner: String,
     @SerializedName("secret") val secret: String,
     @SerializedName("server") val server: String,
-    @SerializedName("farm") val farm: Int,
+    @SerializedName("farm") val farm: Long,
     @SerializedName("title") val title: String,
     @SerializedName("ispublic") val isPublic: Int,
     @SerializedName("isfriend") val isFriend: Int,
